% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent.R
\name{hent_data}
\alias{hent_data}
\title{Get a data file}
\usage{
hent_data(file, full = FALSE, destdir = getwd())
}
\arguments{
\item{file}{string; basename of a file}

\item{full}{logical; whether to save as is or trim the "kapXX_" prefix upon saving.}

\item{destdir}{string; path to the download directory.}
}
\value{
no return value, called for side effects
}
\description{
Downloads a specified file from the \code{data} directory of the source repository. A wrapper for \code{hent_fil()}. The filename must be among those returned by \code{hent_oversikt('data')}.
}
\examples{
\dontrun{
hent_data("kap13_test.bib")

hent_data(hent_oversikt("data")[1], full = TRUE, destdir = tempdir())

datafiler <- hent_oversikt("data")
for (i in datafiler) {hent_data(i)}
}
}
