% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent.R
\name{hent_fil}
\alias{hent_fil}
\title{Get files}
\usage{
hent_fil(file, type, full, destdir)
}
\arguments{
\item{file}{string; passed on from \code{hent_data()} or \code{hent_kode}}

\item{type}{"data" or "kode"; passed on from \code{hent_data()} or \code{hent_kode}}

\item{full}{logical; passed on from \code{hent_data()} or \code{hent_kode}}

\item{destdir}{string; passed on from \code{hent_data()} or \code{hent_kode}}
}
\description{
Helper function for \code{hent_data()} and \code{hent_kode}.
}
\examples{
\dontrun{
hent_fil("kap13_test.bib", "data", full = FALSE, destdir = tempdir())
}
}
