% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent.R
\name{hent_kode}
\alias{hent_kode}
\title{Get a code file}
\usage{
hent_kode(file, full = FALSE, destdir = getwd())
}
\arguments{
\item{file}{string; basename of a file.}

\item{full}{logical; whether to save as is or trim the "kapXX_" prefix upon saving.}

\item{destdir}{string; path to the download directory.}
}
\value{
no return value, called for side effects
}
\description{
Downloads a specified file from the \code{kode} directory of the source repository. A wrapper for \code{hent_fil()}. The filename must be among those returned by \code{hent_oversikt('kode')}.
}
\examples{
\dontrun{
hent_kode("kap05.R")

hent_data(hent_oversikt("kode")[1], full = TRUE, destdir = tempdir())

datafiler <- hent_oversikt("kode")
for (i in datafiler) {hent_kode(i)}
}
}
