% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent.R
\name{hent_oversikt}
\alias{hent_oversikt}
\title{Get list of available files}
\usage{
hent_oversikt(selection = "alle")
}
\arguments{
\item{selection}{string; "alle", "data", or "kode'.}
}
\value{
a character vector with filenames (or filepaths in the case of "alle").
}
\description{
Lists files in source repository available for download with \code{hent_data()} or \code{hent_kode()}.
}
\examples{
\dontrun{
hent_oversikt("data")

hent_oversikt("kode")

hent_oversikt("alle")
}
}
