% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent.R
\name{se_github}
\alias{se_github}
\title{View the source repository on Github}
\usage{
se_github()
}
\description{
Opens a browser at \url{https://github.com/Hegghammer/rbok/}.
}
\examples{
\dontrun{
se_github()
}
}
